requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
redis==5.1.0
google-generativeai>=0.3.1
pydantic>=2.4.2
typing-extensions>=4.8.0
aioredis>=2.0.1 